refactor: unify auth logic in 1 file
